# rimba-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rimba Api",
        "description": "Rimba Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "4db02f0d9ccd97469981e0cb03d1a41e",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "failed operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Adding user",
                "description": "Adding user",
                "operationId": "c70c18636cb91d835fcc23025f48febe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": "Username Here"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "useremail@email.email"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "default": 28
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "failed operation"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user information by id",
                "description": "Returns user data",
                "operationId": "e02f7752f572c91564a2906e469d8259",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "data not found"
                    },
                    "500": {
                        "description": "failed operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user information by id",
                "description": "Update user information by id",
                "operationId": "c0a7289dd73dbc27daecdbffbcc86f9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "default": "Username Here"
                                    },
                                    "email": {
                                        "type": "string",
                                        "default": "useremail@email.email"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "default": 28
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "failed operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user information by id",
                "description": "Delete user data",
                "operationId": "a0ce91d753f85c9a3caee8c136d70e4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "data not found"
                    },
                    "500": {
                        "description": "failed operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ]
}